Fix feature gating
